Moved DBremWatcher to plugins directory
This avoids having the module be registered in multiple shared
libraries.
